# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pizzaroo API",
        "description": "Documentación Swagger para Pizzaroo",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admins": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Obtener todos los administradores",
                "operationId": "27f6447f654b54ee491274c73668fa10",
                "responses": {
                    "200": {
                        "description": "Lista de administradores"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Crear un nuevo administrador",
                "operationId": "e97775d486a8f350753ea15cdf329efb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Administrador creado"
                    }
                }
            }
        },
        "/api/admins/{id}": {
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Actualizar un administrador",
                "operationId": "d1fe95b300f5938f417f9d84f001d509",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Administrador actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Eliminar un administrador",
                "operationId": "4bd429e640a5c6ec94b1dd9c44369866",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Administrador eliminado"
                    }
                }
            }
        },
        "/api/cart-items": {
            "get": {
                "tags": [
                    "Cart Items"
                ],
                "summary": "Obtener todos los ítems del carrito",
                "operationId": "4233993f8a5847bc1b5329e9d1b898a4",
                "responses": {
                    "200": {
                        "description": "Lista de ítems"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cart Items"
                ],
                "summary": "Agregar ítem al carrito",
                "operationId": "0a5df60f7d91780d6e0863dc698a47e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "pizza_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "pizza_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ítem agregado"
                    }
                }
            }
        },
        "/api/cart-items/{id}": {
            "put": {
                "tags": [
                    "Cart Items"
                ],
                "summary": "Actualizar ítem del carrito",
                "operationId": "78a5883c4ab72212dc40f79a6f5f9232",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "pizza_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "pizza_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ítem actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart Items"
                ],
                "summary": "Eliminar ítem del carrito",
                "operationId": "7ee59aafe2ba3f949091fb7d40d907c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ítem eliminado"
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Obtener todas las órdenes",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Crear una nueva orden",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "total",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "total": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada exitosamente"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "put": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Actualizar una orden",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "total": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden actualizada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Eliminar una orden",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden eliminada"
                    }
                }
            }
        },
...
    },
...
}